// src/components/products/types.ts
export type ProductRow = {
    id: string;
    item_name: string;
    item_amount: number;
    item_price: number;
  };
  
  export type CommonStrings = {
    item: string;
    quantity: string;
    unitPrice: string;
    actions: string;
    edit: string;
    delete: string;
    addProduct: string;
    cancel: string;
    save: string;
    total: string;
    emptyText: string;
    searchPlaceholder: string;
  };
  